requirements.txt
aws-cdk-lib
constructs
faker
numpy
pandas
